+++ amplify/backend/function/bridgev2layer/opt/helpers/limiter.js
@@ -1,7 +1,9 @@
-import { rateLimit } from 'express-rate-limit'
+const rateLimit = require('express-rate-limit')
 
-export const limiter = rateLimit({
+const limiter = rateLimit({
 	windowMs: 15 * 60 * 1000, // 15 minutes
 	limit: 100, 
     message: 'Too many requests from this IP, please try again after 15 minutes'
 })
+
+module.exports = limiter;
