+++ amplify/backend/function/bridgev2layer/opt/schema/todoSchema.js
@@ -1,0 +1,11 @@
+const mongoose = require('mongoose');
+
+const todoSchema = new mongoose.Schema({
+    sub: String,
+    title: String,
+    status: { type: String, default: "todo" },
+    created: { type: Date, default: Date.now },
+    finished: Date
+  })
+  
+module.exports = mongoose.model('Todo', todoSchema);
