+++ amplify/backend/function/bridgev2layer/parameters.json
@@ -2,5 +2,5 @@
   "runtimes": [
     "nodejs18.x"
   ],
-  "description": "Updated layer version 2024-07-08T00:29:24.914Z"
+  "description": "Updated layer version 2024-07-09T09:04:14.416Z"
 }
