+++ amplify/backend/function/shortcuthandler/function-parameters.json
@@ -1,6 +1,14 @@
 {
   "permissions": {},
-  "lambdaLayers": [],
+  "lambdaLayers": [
+    {
+      "type": "ProjectLayer",
+      "resourceName": "bridgev2layer",
+      "env": "dev",
+      "version": "Always choose latest version",
+      "isLatestVersionSelected": true
+    }
+  ],
   "environmentVariableList": [
     {
       "cloudFormationParameterName": "database",
