+++ amplify/backend/function/shortcuthandler/src/app.js
@@ -2,11 +2,13 @@ const express = require("express");
 const bodyParser = require("body-parser");
 const awsServerlessExpressMiddleware = require("aws-serverless-express/middleware");
 const mongoose = require("mongoose");
+const shortcutModel = require("/opt/schema/shortcutSchema.js");
+const limiter = require("/opt/helpers/limiter.js");
 // declare a new express app
 const app = express();
 app.use(bodyParser.json());
 app.use(awsServerlessExpressMiddleware.eventContext());
-
+app.use(limiter);
 // Enable CORS for all methods
 app.use(function (req, res, next) {
   res.header("Access-Control-Allow-Origin", "*");
@@ -15,18 +17,6 @@ app.use(function (req, res, next) {
 });
 
 mongoose.connect(process.env.DATABASE);
-
-const shortcutSchema = mongoose.Schema({
-  sub: String,
-  title: String,
-  url: String,
-  createdBy: {
-    type: Date,
-    default: Date.now(),
-  },
-});
-
-const shortcutModel = mongoose.model("shortcut", shortcutSchema);
 
 app.get("/shortcut", async function (req, res) {
   const query = req.query;
