+++ amplify/backend/function/shortcuthandler/src/package.json
@@ -7,8 +7,7 @@
   "dependencies": {
     "aws-serverless-express": "^3.3.5",
     "body-parser": "^1.17.1",
-    "express": "^4.15.2",
-    "mongoose": "^8.1.0"
+    "express": "^4.15.2"
   },
   "devDependencies": {
     "@types/aws-lambda": "^8.10.92"
