+++ amplify/backend/function/todohandler/function-parameters.json
@@ -1,5 +1,13 @@
 {
-  "lambdaLayers": [],
+  "lambdaLayers": [
+    {
+      "type": "ProjectLayer",
+      "resourceName": "bridgev2layer",
+      "env": "dev",
+      "version": "Always choose latest version",
+      "isLatestVersionSelected": true
+    }
+  ],
   "environmentVariableList": [
     {
       "cloudFormationParameterName": "database",
