+++ amplify/backend/function/todohandler/todohandler-cloudformation-template.json
@@ -16,6 +16,10 @@
     "s3Key": {
       "Type": "String"
     },
+    "functionbridgev2layerArn": {
+      "Type": "String",
+      "Default": "functionbridgev2layerArn"
+    },
     "database": {
       "Type": "String"
     }
@@ -85,7 +89,11 @@
           ]
         },
         "Runtime": "nodejs18.x",
-        "Layers": [],
+        "Layers": [
+          {
+            "Ref": "functionbridgev2layerArn"
+          }
+        ],
         "Timeout": 25
       }
     },
