+++ amplify/backend/function/userhandler/src/app.js
@@ -2,10 +2,13 @@ const express = require('express')
 const bodyParser = require('body-parser')
 const awsServerlessExpressMiddleware = require('aws-serverless-express/middleware')
 const mongoose = require('mongoose')
+const limiter = require('/opt/helpers/limiter.js')
+
 // declare a new express app
 const app = express()
 app.use(bodyParser.json())
 app.use(awsServerlessExpressMiddleware.eventContext())
+app.use(limiter)
 
 // Enable CORS for all methods
 app.use(function(req, res, next) {
